StyleEditor - fix Marker icon options
After creating a polyline (route) you can not create a Marker anymore.
The current element in styleeditor is the actual polyline and by
calling getDefaultIcon() styleditor will crash (as it assumes
current elements are always markers and does not check them properly).

Workaround: Try to default iconOptions but catch exception if not
available. This should be implemented properly upstream.